// src/app/demo/models/modelsplanilla/categoriaviaticoviewmodel.ts
export class categoriaviatico{
    cavi_Id?: number;
    cavi_Descripcion?: string;
    usua_Creacion?: string;
    cavi_FechaCreacion?: string;
    usua_Modificacion?: string;
    cavi_FechaModificacion?: string;
    cavi_Estado?: boolean;
    codigo?: number;
    usuarioCreacion?:string;
    usuarioModificacion?:string;
}